for test

+++ merge.py
@@ -32,26 +32,26 @@ class run_process_merge():
         else:
             raise ValueError("File not found !!")
         
-        df_int_mapp = self.source_int_mapp(sheet1)
+        # df_int_mapp = self.source_int_mapp(sheet1)
         df_sys_name = self.source_pl_config(sheet1)
-        df_table_def = self.source_table_def(sheet1)
+        # df_table_def = self.source_table_def(sheet1)
         df_ddl = self.source_ddl(sheet2)
         
-        for mvp in self.str_mvp:
-            sum_df = pandas.concat([df_table_def[mvp], df_int_mapp[mvp], df_sys_name[mvp]], axis=0, ignore_index=True)
-            sum_df['Note UAT Deploy Date'] = self.date
-            sum_df['Git_Path'] = sum_df['Storage_Path'].apply(lambda x: "{}/{}/{}/".format(self.date_fmt, x, mvp))
-            sum_df["Path"] = sum_df[["Storage_Path", "File_Name"]].apply(lambda x: "/".join(x), axis =1)
-            sum_df["Full_Path"] = sum_df[["Git_Path", "File_Name"]].apply(lambda x: "".join(x), axis =1)
-            sum_df["Obsolete"] = ""
-            sum_df["Checklist"] = sum_df[["Storage", "Container", "Full_Path","Path"]].apply(lambda x: ",".join(x), axis =1)  
+        # for mvp in self.str_mvp:
+        #     sum_df = pandas.concat([df_table_def[mvp], df_int_mapp[mvp], df_sys_name[mvp]], axis=0, ignore_index=True)
+        #     sum_df['Note UAT Deploy Date'] = self.date
+        #     sum_df['Git_Path'] = sum_df['Storage_Path'].apply(lambda x: "{}/{}/{}/".format(self.date_fmt, x, mvp))
+        #     sum_df["Path"] = sum_df[["Storage_Path", "File_Name"]].apply(lambda x: "/".join(x), axis =1)
+        #     sum_df["Full_Path"] = sum_df[["Git_Path", "File_Name"]].apply(lambda x: "".join(x), axis =1)
+        #     sum_df["Obsolete"] = ""
+        #     sum_df["Checklist"] = sum_df[["Storage", "Container", "Full_Path","Path"]].apply(lambda x: ",".join(x), axis =1)  
             
-            df_new = sum_df.loc[:, ['Storage', 'Container', 'Git_Path', 'Storage_Path','File_Name', 'Note UAT Deploy Date', 'Obsolete', 'Checklist']]
-            if df_new.empty is False:
-                self.write_from_source(df_new, mvp)
-                print(f"{mvp} files count: {len(df_new)} rows and write to excel completed.")
-            else:
-                raise Exception("Dataframe is empty !!")
+        #     df_new = sum_df.loc[:, ['Storage', 'Container', 'Git_Path', 'Storage_Path','File_Name', 'Note UAT Deploy Date', 'Obsolete', 'Checklist']]
+        #     if df_new.empty is False:
+        #         self.write_from_source(df_new, mvp)
+        #         print(f"{mvp} files count: {len(df_new)} rows and write to excel completed.")
+        #     else:
+        #         raise Exception("Dataframe is empty !!")
             
     def write_from_source(self, df, mvp):
         
@@ -197,6 +197,7 @@ class run_process_merge():
         
         return dict_df
     
+    
     def source_ddl(self, dataframe):
         
         print("================= ddl ===================")
@@ -204,8 +205,8 @@ class run_process_merge():
 
         if new_df.empty is False:
             new_col = new_df['VIEW_TABLE'].map(lambda x: str(x)[2:]).str.split(".", n=1, expand=True)
-            new_df["Folder"]= new_col[0]
-            new_df["File"]= new_col[1]
+            new_df["Folder"] = new_col[0]
+            new_df["File"] = new_col[1]
             print(f"count file ddl: {len(new_df)} files")
             
             mvp1 = new_df[new_df['MVP'] == 'MVP1']
@@ -214,48 +215,48 @@ class run_process_merge():
             mvp4 = new_df[new_df['MVP'] == 'MVP4']
             mvp6 = new_df[new_df['MVP'] == 'MVP6']
             all_mvp = [mvp1,mvp2,mvp3,mvp4,mvp6]
-            
             parent_dir = os.getcwd() + f'/filename/DDL/{self.date}'
             
             for str_mvp, all_mvp in zip(self.str_mvp, all_mvp):
-                destination  = os.path.join(parent_dir, str_mvp)
-                for fol, file in zip(all_mvp['Folder'].values.tolist(), all_mvp['File'].values.tolist()):
-                    path_in = f'{parent_dir}/VIEW/{fol}'
-                    path_out = f'{destination}/{fol}'
-                    os.makedirs(path_out, exist_ok=True)
-                    full_name = str(file) + '.sql'
+                if all_mvp.empty is False:
+                    destination  = os.path.join(parent_dir, str_mvp)
+                    for fol, file in zip(all_mvp["Folder"].values.tolist(), all_mvp["File"].values.tolist()):
+                        path_in = f'{parent_dir}/VIEW/{fol}'
+                        path_out = f'{destination}/{fol}'
+                        os.makedirs(path_out, exist_ok=True)
+                        full_name = str(file) + '.sql'
+                        
+                        if os.path.isfile(os.path.join(path_in, full_name)):
+                            os.remove(os.path.join(path_out, full_name))
+                            shutil.copy(os.path.join(path_in, full_name), path_out)
+                        else:
+                            print(f"file: {full_name} does not exist on folder: {fol}, '{str_mvp}'")
+                
+                    # set dataframe
+                    sum_df = pandas.DataFrame(zip(all_mvp["File"].values.tolist(), all_mvp["Folder"].values.tolist()), columns=['File', 'Folder']).reset_index(drop=True)
+                    sum_df['File_Name'] = sum_df['File'].apply(lambda x: str(x).upper() + '.sql')
+                    sum_df['Note UAT Deploy Date'] = self.date
+                    sum_df['Storage'] = self.storage
+                    sum_df['Container'] = self.container
+                    sum_df['Storage_Path'] = sum_df[["Folder", "File_Name"]].apply(lambda x: "/".join(x), axis=1)
+                    sum_df["Git_Path"] =  sum_df['Storage_Path'].apply(lambda x: "VIEWS/{}".format(x)) 
+                    sum_df["Checklist"] =  sum_df['Storage_Path'].apply(lambda x: "ddl_script_replace/process_migration/{}".format(x)) 
+                    df_new = sum_df.loc[:, ['Storage', 'Container', 'Git_Path', 'Note UAT Deploy Date', 'Checklist']]
+                
+                    # write to sheet 
+                    print(f"{str_mvp} ddl files count: {len(df_new)} rows and write to excel completed.")
+                    sheet_name = f'Checklist_DDL_{str_mvp}'
+                    self.sheet = self.wb.create_sheet(sheet_name)
+                    self.sheet.title = sheet_name
+                    rows = dataframe_to_rows(df_new, header=True, index=False)
+                    for r_idx, row in enumerate(rows, 1):
+                        for c_idx, val in enumerate(row, 1):
+                            value = self.sheet.cell(row=r_idx, column=c_idx)
+                            value.value = val
                     
-                    if os.path.isfile(os.path.join(path_in, full_name)):
-                        os.remove(os.path.join(path_out, full_name))
-                        shutil.copy(os.path.join(path_in, full_name), path_out)
-                    else:
-                        print(f"file: {full_name} does not exist on folder: {fol}, '{str_mvp}'")
-                
-                # set dataframe
-                sum_df = pandas.DataFrame(zip(all_mvp['File'].values.tolist(), all_mvp['Folder'].values.tolist()), columns=['File', 'Folder']).reset_index(drop=True)
-                sum_df['File_Name'] = sum_df['File'].apply(lambda x: str(x).upper() + '.sql')
-                sum_df['Note UAT Deploy Date'] = self.date
-                sum_df['Storage'] = self.storage
-                sum_df['Container'] = self.container
-                sum_df['Storage_Path'] = sum_df[["Folder", "File_Name"]].apply(lambda x: "/".join(x), axis=1)
-                sum_df["Git_Path"] =  sum_df['Storage_Path'].apply(lambda x: "VIEWS/{}".format(x)) 
-                sum_df["Checklist"] =  sum_df['Storage_Path'].apply(lambda x: "ddl_script_replace/process_migration/{}".format(x)) 
-                df_new = sum_df.loc[:, ['Storage', 'Container', 'Git_Path', 'Note UAT Deploy Date', 'Checklist']]
-                
-                # write to sheet 
-                print(f"{str_mvp} ddl files count: {len(df_new)} rows and write to excel completed.")
-                sheet_name = f'Checklist_DDL_{str_mvp}'
-                self.sheet = self.wb.create_sheet(sheet_name)
-                self.sheet.title = sheet_name
-                rows = dataframe_to_rows(df_new, header=True, index=False)
-                for r_idx, row in enumerate(rows, 1):
-                    for c_idx, val in enumerate(row, 1):
-                        value = self.sheet.cell(row=r_idx, column=c_idx)
-                        value.value = val
-                
-                self.file_name = f'{self.path}/deployment_checklist_{self.date_fmt}.xlsx'
-                self.wb.save(self.file_name)
-        else:
-            print("count file ddl: empty files")
+                    self.file_name = f'{self.path}/deployment_checklist_{self.date_fmt}.xlsx'
+                    self.wb.save(self.file_name)
+                else:
+                    print(f"{str_mvp} ddl files is empty")
             
         print("=========================================")
